requirements.txt
aiogram>=3.17,<4
openai>=1.60,<2
python-dotenv>=1.0,<2
pdfplumber>=0.11,<0.12
python-docx>=1.1,<2
